Move the global Jobmanager to the main application class. This forced us to weaken access to the apis in the util module

+++ utils/src/main/java/com/pairapp/util/TaskManager.java
@@ -1,19 +1,12 @@
 package com.pairapp.util;
 
-import android.app.Application;
 import android.os.Handler;
 import android.os.Looper;
 import android.support.annotation.NonNull;
 
-import com.path.android.jobqueue.JobManager;
-import com.path.android.jobqueue.TagConstraint;
-import com.path.android.jobqueue.config.Configuration;
-import com.path.android.jobqueue.di.DependencyInjector;
-
 import java.io.File;
 import java.io.IOException;
-import java.util.Timer;
-import java.util.TimerTask;
+import java.lang.ref.WeakReference;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
@@ -55,36 +48,25 @@ public class TaskManager {
                     = new ThreadPoolExecutor(CORE_POOL_SIZE, MAXIMUM_POOL_SIZE, KEEP_ALIVE,
                     TimeUnit.SECONDS, networkWorkQueue);
 
-    private static JobManager jobManager;
+    private static WeakReference<JobRunner> jobManager;
 
 
-    public static void init(final Application application, DependencyInjector injector) {
+    public static void init(JobRunner runner) {
         if (!initialised.getAndSet(true)) {
             PLog.w(TAG, "initialising %s", TAG);
-            final Configuration config = new Configuration.Builder(application)
-                    .injector(injector)
-                    .customLogger(new PLog("JobManager"))
-                    .jobSerializer(new Task.JobSerializer())
-                    .build();
-            final Timer timer = new Timer("cleanup Timer", true);
-            timer.schedule(new TimerTask() {
-                @Override
-                public void run() {
-                    try {
-                        jobManager = new JobManager(application, config);
-                        jobManager.start();
-                        File file = Config.getTempDir();
-                        if (file.exists()) {
-                            org.apache.commons.io.FileUtils.cleanDirectory(Config.getTempDir());
-                        }
-                    } catch (IOException ignored) {
+            jobManager = new WeakReference<>(runner);
+            try {
+                JobRunner jobRunner = jobManager.get();
+                if (jobRunner == null) throw new IllegalStateException();
+                jobRunner.start();
 
-                    } finally {
-                        timer.cancel();
-                    }
+                File file = Config.getTempDir();
+                if (file.exists()) {
+                    org.apache.commons.io.FileUtils.cleanDirectory(Config.getTempDir());
                 }
-            }, 1); //immediately
+            } catch (IOException ignored) {
 
+            }
         }
     }
 
@@ -93,7 +75,11 @@ public class TaskManager {
         if (job == null || !job.isValid()) {
             throw new IllegalArgumentException("invalid job");
         }
-        return jobManager.addJob(job);
+        JobRunner jobRunner = jobManager.get();
+        if(jobRunner == null){
+            throw new IllegalStateException();
+        }
+        return jobRunner.runJob(job);
     }
 
     public static void executeOnMainThread(Runnable r) {
@@ -133,7 +119,7 @@ public class TaskManager {
             }
             expressQueueLength++;
         }
-            cachedThreadPool.execute(runnable);
+        cachedThreadPool.execute(runnable);
     }
 
     public static Future<?> executeNow(Callable<?> callable, boolean requiresNetwork) {
@@ -160,7 +146,11 @@ public class TaskManager {
     }
 
     public static void cancelJobSync(String tag) {
-        jobManager.cancelJobs(TagConstraint.ALL, tag);
+        JobRunner jobRunner = jobManager.get();
+        if(jobRunner == null){
+            throw new IllegalStateException();
+        }
+        jobRunner.cancelJobs(tag);
     }
 
     private static class SmartThread extends Thread {
@@ -180,4 +170,12 @@ public class TaskManager {
     }
 
     private static final Object expressQueueLock = new Object();
+
+    public interface JobRunner {
+        long runJob(Task task);
+
+        void cancelJobs(String tag);
+
+        void start();
+    }
 }
